// src/service-api/generated/models/activitiesActivityTypeDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type ActivitiesActivityTypeDto = typeof ActivitiesActivityTypeDto[keyof typeof ActivitiesActivityTypeDto];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActivitiesActivityTypeDto = {
  CALL: 'CALL',
  EMAIL: 'EMAIL',
  MEETING: 'MEETING',
  LINKEDIN: 'LINKEDIN',
  NOTE: 'NOTE',
} as const;