// src/service-api/generated/models/dealsUpdateRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */
import type { DealsDealStageDto } from './dealsDealStageDto';
import type { DealsUpdateRequestValue } from './dealsUpdateRequestValue';

export interface DealsUpdateRequest {
  /** @nullable */
  contactId?: string | null;
  /** @nullable */
  closeDate?: string | null;
  stage?: DealsDealStageDto;
  value?: DealsUpdateRequestValue;
  title?: string;
}