// src/service-api/generated/models/dealsTaskDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */
import type { DealsTaskDtoOutcome } from './dealsTaskDtoOutcome';
import type { DealsTaskStatusDto } from './dealsTaskStatusDto';
import type { DealsTaskTypeDto } from './dealsTaskTypeDto';

export interface DealsTaskDto {
  updatedAt: string;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  outcome: DealsTaskDtoOutcome;
  completed: boolean;
  /** @nullable */
  dueDate: string | null;
  title: string;
  status: DealsTaskStatusDto;
  type: DealsTaskTypeDto;
  /** @nullable */
  assignedTo: number | null;
  /** @nullable */
  companyId: string | null;
  /** @nullable */
  dealId: string | null;
  id: string;
}