// src/service-api/generated/models/updateTaskRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */
import type { TaskTypeDto } from './taskTypeDto';

export interface UpdateTaskRequest {
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  dueDate?: string | null;
  type?: TaskTypeDto;
  title?: string;
}